docker/content-moderator/requirements.txt
Flask==2.3.3
Pillow==10.0.0
python-magic==0.4.27
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0